It is better use try catch finally block
I notice on automation if exception occur. It stuck with exception and it become more hard to handle if it is integrated with other application. My recommendation while using it. Please follow below manner. Using finally block to phantom.exit() should be mandatory like best case practice else each thread create an instance of phantomjs.exe and you have to go to Task manger and kill the process.